requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.2.0
Brotli==1.0.9
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
discord.py==1.5.1
discord.py[voice]
colorlog
asyncio
fonttools==4.17.1
fs==2.4.11
idna==2.10
lxml==4.6.2
multidict==4.7.6
Pillow==8.0.1
pycparser==2.20
PyNaCl
pip
wheel
pytz==2020.4
requests==2.25.0
six==1.15.0
unicodedata2==13.0.0.post2
urllib3==1.26.2
utils==1.0.1
yarl==1.5.1
zopfli==0.1.7
